pyproject.toml
[project]
name = "x-ray"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "streamlit>=1.51.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
